refactor: merged  subpackage _legacy_services_wrapper into _legacy_subpackages and reduce the public interface of the latter
