--- rats-processors/src/python/rats/processors/_legacy_services_wrapper/__init__.py
@@ -1,15 +1,0 @@
-"""Exposes selected services from rats.processors._legacy_subpackages."""
-
-from ._container import (
-    LegacyServicesWrapperContainer,
-    Services,
-)
-from ._interfaces import IPipelineRunner, IPipelineRunnerFactory, IPipelineToDot
-
-__all__ = [
-    "IPipelineRunner",
-    "IPipelineRunnerFactory",
-    "IPipelineToDot",
-    "LegacyServicesWrapperContainer",
-    "Services",
-]
