+++ rats-processors/src/python/rats/processors/_legacy_subpackages/__init__.py
@@ -1,16 +1,21 @@
-from . import dag, registry, utils, ux
-from .dag import display_dag
-from .training import ScatterGatherBuilders, TrainAndEvalBuilders
-from .ux import Pipeline, PipelineBuilder
+# This defines the interface of _legacy_subpackages for use by the rest of rats.processors.
+
+
+from . import _typing as typing
+from ._container import (
+    LegacyServicesWrapperContainer,
+    Services,
+)
+from ._interfaces import IPipelineRunner, IPipelineRunnerFactory, IPipelineToDot
+from .ux import CombinedPipeline, Task
 
 __all__ = [
-    "PipelineBuilder",
-    "display_dag",
-    "TrainAndEvalBuilders",
-    "ScatterGatherBuilders",
-    "Pipeline",
-    "dag",
-    "ux",
-    "utils",
-    "registry",
+    "Task",
+    "CombinedPipeline",
+    "typing",
+    "IPipelineRunner",
+    "IPipelineRunnerFactory",
+    "IPipelineToDot",
+    "LegacyServicesWrapperContainer",
+    "Services",
 ]
