+++ rats-processors/src/python/rats/processors/_legacy_subpackages/_container.py
@@ -1,7 +1,7 @@
 from rats import apps
 from rats.app import RatsApp as LegacyApp
-from rats.processors._legacy_subpackages import dag, ux
 
+from . import dag, ux
 from ._interfaces import IPipelineRunnerFactory, IPipelineToDot
